create UserRepository and UserRepositoryFactory for database access
